Refactored code in Sidebar and AppSwitcher
