+++ lib/components/layouts/AppSwitcher/index.js
@@ -2,27 +2,18 @@ import React, { useRef } from "react";
 import { AnimatePresence, motion } from "framer-motion";
 import { useHotkeys } from "react-hotkeys-hook";
 import classnames from "classnames";
-import PropTypes from "prop-types";
 
 import Body from "./Body";
 import { Portal, Backdrop } from "../../../atoms";
 import { useOnClickOutside } from "../../../utils";
 import {
-  SWITCHER_SIZE,
-  BLUR_INITIAL,
-  BLUR_FINAL,
   TRANSITION,
-} from "../../../constants/overlay";
-
-const sizes = { lg: "lg", sm: "sm" };
+  ANIMATION_VARIANTS,
+  APPSWITCHER_PROPTYPES,
+  SIZES,
+} from "./constant";
 
 const noop = () => {};
-
-const envs = {
-  development: "development",
-  staging: "staging",
-  production: "production",
-};
 
 const getSubdomain = () => {
   let host = window.location.host;
@@ -34,21 +25,9 @@ const getSubdomain = () => {
   return subdomain;
 };
 
-const animationVariants = {
-  open: {
-    x: 0,
-    filter: BLUR_INITIAL,
-  },
-  closed: {
-    x: -SWITCHER_SIZE,
-    filter: BLUR_FINAL,
-  },
-};
-
 const AppSwitcher = ({
   size = "sm",
   isOpen,
-  onClose = noop,
   loading = false,
   className = "",
   closeOnEsc = true,
@@ -59,6 +38,7 @@ const AppSwitcher = ({
   neetoApps = [],
   recentApps = [],
   isSidebarOpen = false,
+  onClose = noop,
   ...otherProps
 }) => {
   const paneWrapper = useRef();
@@ -84,7 +64,7 @@ const AppSwitcher = ({
             })}
           >
             <motion.div
-              variants={animationVariants}
+              variants={ANIMATION_VARIANTS}
               initial="closed"
               animate="open"
               exit="closed"
@@ -96,8 +76,8 @@ const AppSwitcher = ({
               key="switcher-wrapper"
               data-cy="switcher-wrapper"
               className={classnames("neeto-ui-app-switcher__wrapper", {
-                "neeto-ui-app-switcher__wrapper-size--sm": size === sizes.sm,
-                "neeto-ui-app-switcher__wrapper-size--lg": size === sizes.lg,
+                "neeto-ui-app-switcher__wrapper-size--sm": size === SIZES.sm,
+                "neeto-ui-app-switcher__wrapper-size--lg": size === SIZES.lg,
                 [className]: className,
               })}
               {...otherProps}
@@ -122,40 +102,5 @@ const AppSwitcher = ({
   );
 };
 
-AppSwitcher.propTypes = {
-  /**
-   * Size of the AppSwitcher
-   */
-  size: PropTypes.oneOf(Object.values(sizes)),
-  isOpen: PropTypes.bool,
-  onClose: PropTypes.func,
-  loading: PropTypes.bool,
-  className: PropTypes.string,
-  /**
-   * To close AppSwitcher on esc key press
-   */
-  closeOnEsc: PropTypes.bool,
-  /**
-   * To close AppSwitcher on outside click
-   */
-  closeOnOutsideClick: PropTypes.bool,
-  /**
-   * Name of the active application
-   */
-  activeApp: PropTypes.string,
-  /**
-   * Organization subdomain name
-   */
-  subdomain: PropTypes.string,
-  /**
-   * List of names of enabled neetoApps
-   */
-  neetoApps: PropTypes.arrayOf(PropTypes.string),
-  /**
-   * List of names of recently created neetoApps
-   */
-  recentApps: PropTypes.arrayOf(PropTypes.string),
-  environment: PropTypes.oneOf(Object.values(envs)).isRequired,
-};
-
+AppSwitcher.propTypes = APPSWITCHER_PROPTYPES;
 export default AppSwitcher;
